requirements.txt
glob2
pandas
tensorflow
pillow
lxml
object-detection